This PR replaces `serde_json` by `json_deserializer`, improving the performance of reading json strings by 2x and boolean by -30%, and +10% for floats:

```
read i32 2^20           time:   [68.065 ms 68.256 ms 68.444 ms]                          
                        change: [-0.8660% +1.5241% +3.8068%] (p = 0.20 > 0.05)

read f64 2^20           time:   [90.453 ms 90.682 ms 90.904 ms]                          
                        change: [+10.496% +10.879% +11.269%] (p = 0.00 < 0.05)

read utf8 2^20          time:   [51.573 ms 51.699 ms 51.825 ms]                           
                        change: [-56.883% -56.714% -56.550%] (p = 0.00 < 0.05)

read bool 2^20          time:   [33.284 ms 33.483 ms 33.837 ms]                           
                        change: [-30.126% -29.623% -28.852%] (p = 0.00 < 0.05)
```

Note that there is an API change: we now require to pass `json_deserializer::Value`, not `serde_json::Value`, to the deserializer. `json_deserializer` is exposed under `arrow2::io::json::read::json_deserializer` for convenience / version matching.